Refactor solver logic into separate modules
- Move all solver strategies from Field to sudoku/solver package - Add
individual modules for each strategy (e.g. naked_pairs, x_wing) -
Implement shared generator utilities in solver/utils.py - Update imports
and usage in pg_gui/board.py - Update __init__.py files and exports
accordingly - Clean up unused imports and code in affected files
