+++ Justfile
@@ -1,0 +1,2 @@
+run:
+    PYTHONTRACEMALLOC=1 PYTHONASYNCIODEBUG=1 uv run python -m sudoku
