+++ sudoku/__init__.py
@@ -1,3 +1,7 @@
-from . import cell, field, types
+# from . import cell, field, types
+from .action import Action
+from .cell import Cell
+from .field import Field
+from .types import CellPosition, CellValue
 
-__all__ = ["cell", "field", "types"]
+__all__ = ["Action", "Cell", "Field", "CellPosition", "CellValue"]
